third_party/WebKit/LayoutTests/http/tests/security/xss-DENIED-window-name-alert-expected.txt

CONSOLE MESSAGE: line 6: FAIL
CONSOLE MESSAGE: line 7: compatible with old versions = true
